Added boilerplate generator
This is an absolutely massive commit which refactors all of the deploy
code into a generator binary that can create boilerplate projects.

All of the new generator code is currently missing documentation and
tests and is failing lint, but is in a working state based on manual
verification. Little by little...